check coverage without hsm tests

+++ cmd/microservice/main.go
@@ -13,6 +13,7 @@ import (
 	"net/http"
 	"net/url"
 	"os"
+	"path/filepath"
 	"plugin"
 	"strconv"
 	"strings"
@@ -97,6 +98,11 @@ func newHSMContext() (*hsmContext, error) {
 }
 
 func destroyHSMContext(hc *hsmContext) {
+	if hc == nil {
+		slog.Error("destroyHSMContext error, input param is nil")
+		return
+	}
+
 	defer hc.ctx.Destroy()
 	err := hc.ctx.Finalize()
 	if err != nil {
@@ -134,6 +140,11 @@ func newHSMSession(hc *hsmContext) (*hsmSession, error) {
 }
 
 func destroyHSMSession(hs *hsmSession) {
+	if hs == nil {
+		slog.Error("destroyHSMSession err, input param is nil")
+		return
+	}
+	fmt.Println("=======", hs.session)
 	err := hs.c.ctx.CloseSession(hs.session)
 	if err != nil {
 		slog.Error("pkcs11 error closing session", "err", err)
@@ -174,7 +185,15 @@ func loadAuditHook() func(f http.Handler) http.Handler {
 		}
 	}
 
-	plug, err := plugin.Open("plugins/audit_hooks.so")
+	testDir, err := os.Getwd()
+	if err != nil {
+		panic(err)
+	}
+	pluginPath := filepath.Join(testDir, "..", "..", "plugins", "audit_hooks.so")
+
+	slog.Info("Install plugin", "path", pluginPath)
+
+	plug, err := plugin.Open(pluginPath)
 	if err != nil {
 		panic(err)
 	}
